changed representation from doubles to floats
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18268 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 5d093ead18953cb434dae3e1de2e365714b82d8b

+++ sandbox/descriptors_3d/include/descriptors_3d/descriptors_3d.h
@@ -63,7 +63,7 @@ class Descriptor3D
     }
     ;
 
-    virtual bool compute(Eigen::MatrixXd** result, bool debug) = 0;
+    virtual bool compute(Eigen::MatrixXf** result, bool debug) = 0;
 
     void setData(const robot_msgs::PointCloud* data, cloud_kdtree::KdTree* data_kdtree)
     {
@@ -106,7 +106,7 @@ class LocalGeometry: public Descriptor3D
     {
     }
 
-    virtual bool compute(Eigen::MatrixXd** result, bool debug);
+    virtual bool compute(Eigen::MatrixXf** result, bool debug);
 
     virtual bool readyToCompute();
